clean slide dom method, rename file
